Increase GPU power buffer sizes
Instead of using Nc as the batch size, we now use Nc*Ns as the batch
size.  IOW, we now setup the batched plan to perform all FFTs on the
input buffer with one execut call (per output product).  This makes way
better use of the GPU (i.e. it's way faster), but we need a post-FFT
kernel to perform the integrations.  We will also need to increase the
host output power buffer sizes.