Eliminate "accumulate" CuFFT callback
The distinction between "store" and "accumulate" CuFFT callbacks is only
for single polarization input and the benefit of such distinction in
marginal, so for now we are eliminating the concept of an "accumulate"
CuFFT callback (and making the "store" callback actually accumulate).

There may come a time when the effort to distinguish between "store" and
"accumulate" is worthwhile, but we're not there yet.