Fix count of spectra in call to accumulate kernel
The call to the accumulate kernel used to specify that Na spectra within
the input buffer should be accumulated together.  This is obviously the
wrong value to use when Na specifies more spectra than fit in the input
buffer!  The correct number to use is the minimum of Na and Ns.