Add socket output functionality
The output packets are jumbo packets and should be sent over an
interface with an appropriate MTU (at least 8600, preferably 9000).

The outbound interface needs to have a large enough txqueuelen setting
or SndbufErrors will result (i.e. outbound packets will be dropped
before transmission).  The SndbufErrors counter can be seen (when
non-zero) in the output of the `netstat -s -u` command.  The Linux
(Ubuntu 12.04) txqueuelen default of 1,000 is not large enough, but
increasing to 10,000 prevented SndbufErrors from incrementing during
testing.