Add commented-out code to set socket options
This adds code to set scoket options after the output buffer sizes are
known.  This code was added during development in an attempt to
eliminate packet loss on the transmit side by increasing the SO_SNDBUF
size.  Ultimately it was determined that this was not the cause of
transmit side packet loss so this code has been commented out so it can
be used for other purposes, if needed.  The transmit packet loss was
actually caused by too low of a txqueuelen setting.