Add preliminary net throttle code
When multiple nodes are dumping to a single network destination, it is
possible that simultaneous bursts of packets from multiple nodes will
overwhelm the packet buffers in the switch and lead to packet loss.  To
help prevent that, the senders should limit their output data rates.
Preliminary code to do that is added here, though it is not in the final
form needed to prevent packet loss in the switch with multiple senders.