Change -r/--rate option to be absolute Gbps limit
We really care about limiting the output network data rate to a specific
value so that multiple senders will not flood the switch with more data
than can be buffered for a single destination.

This version assumes that the output threads will be outputting
simultaneously in parallel so it carefully calculates different output
rates for the various output products based on their data volume.  This
is done so to avoid having each of the N threads outputting at the
desired rate resulting in an output rate that is N times the desired
rate.

In practice, it is not clear that things are working as expected.  Need
to verify measured/calculated performance at receiver with statistics
measured/calculated by rawspec.