Pad fb pkt header for optimal packet socket rx
To ensure that the data portion of the filterbank packet payloads are
aligned to 512 byte boundaries, the filterbank header of the packet is
padded.  The calculation of the padding is done specifically for
receivers using packet sockets.  The padding will be suboptimal for
receivers using the standard `recv()` family of system calls (but those
are alreayd suboptimal compared to packet sockets anyway).