Full pol/stokes modes now input conjugation aware
The sign of the imaginary component of the cross-pol product and the
closely related Stokes V parameter is dependent upon the input
conjugation of the data.  If the input data are conjugated, for example
do to an odd number of lower sidband mixes in the downconversion
process, then the sign will be the opposite of what it would be for
un-conjugated input data.

The rawspec context structure now has a `input_conjugated` field that
indicates the input data are conjugated (e.g. the frequencies are
flipped).  The rawspec program sets this field to 1 if OBSBW is negative
(indicating a spectral flip in the input), 0 otherwise.

There are many ways to get the sign wrong, so it's quite likely this
will not be the last word on this topic!  At least this has been shown
to produce full-pol filterbank files that dspsr can convert to PSRCHIVE
files that have the same sign for circular polarization (Stokes V) as
dspsr produces from the same RAW input files (at least for S band data
of B0329+54 from GBT).