Fix texture object access from load callback
The load_callback function needs to be able to access the texture object
from device memory.  Trying to access it in host memory from within the
callback is ill-fated (to put it kindly).  For now, we put it at global
scope (like the old texture reference was) and use cudaMemcpyToSymbol to
copy the texture object to device memory after it has been created in
host memory.