Fix case NPOL = 2
The case where NPOL = 2 in the GUPPI header (i.e., when the data
is single polarization) was not handled correctly. The number of
polarizations would be interpreted as 2 when reading the GUPPI header,
causing problems later on.

Here we set the number of polarizations to 1 when NPOL = 2.
Additionally, we fail if NPOL is not either 2 or 4, since in
rawspec_initialize() it is checked that the number of polarizations
is either 1 or 2. However, an NPOL = 1 in the header (which is
meaningless) would get through and be interpreted as 1 polarization
unless we fail in this case.